requirements.txt
requests
browser_cookie3
brotlipy
bs4
